requirements.txt
ipython==7.7.0
ipython-genutils==0.2.0
MarkupSafe==1.1.1
numpy==1.17.0
pandas==0.25.0
python-docx==0.8.10
xlwt==1.3.0
